Move utility functions to own module
